FIX: draw_idle works OK with the RE
If we are running these callbacks _from the RE_ we have a periodic
re-draw of all figures know to pyplot.

If we are explicitly calling `draw` from the callback (hence from the
background thread) then we have a race between the two draws and end up getting
either the start of the render process (due to grabbing and shipping a
half-rendered image to front end) or the end (due to calling a clear in the
middle of a draw!).